Auth

+++ MenrvaBooks/src/main/kotlin/com/menrva/repositories/TagRepository.kt
@@ -9,6 +9,6 @@ import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TagRepository : JpaRepository<Tag, Int> {
+interface TagRepository : JpaRepository<Tag, Long> {
     fun findByName(name: String): List<Tag>
 }
